+++ qoqo/tests/integration/operations/define_operations.rs
@@ -10,9 +10,13 @@
 // express or implied. See the License for the specific language governing permissions and
 // limitations under the License.
 
+#[cfg(feature = "unstable_operation_definition")]
+use super::pragma_operations::new_circuit;
 use pyo3::prelude::*;
 use qoqo::operations::*;
 use roqoqo::operations::*;
+#[cfg(feature = "unstable_operation_definition")]
+use roqoqo::Circuit;
 #[cfg(feature = "json_schema")]
 use roqoqo::ROQOQO_VERSION;
 use std::collections::{HashMap, HashSet};
@@ -241,7 +245,65 @@ fn test_pyo3_new_input_bit() {
     })
 }
 
-/// Test DefinitionFloat, DefinitionComplex, DefinitionUsize, DefinitionBit, InputSymbolic name() function/input
+/// Test GateDefinition new() function
+#[test]
+#[cfg(feature = "unstable_operation_definition")]
+fn test_pyo3_new_gate_definition() {
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation = py.get_type::<GateDefinitionWrapper>();
+        let new_op = operation
+            .call1((
+                new_circuit(py),
+                String::from("ro"),
+                vec![1],
+                vec!["a".to_owned(), "b".to_owned()],
+            ))
+            .unwrap()
+            .downcast::<PyCell<GateDefinitionWrapper>>()
+            .unwrap();
+
+        let input_definition = Operation::from(GateDefinition::new(
+            Circuit::new(),
+            "ro".into(),
+            vec![1],
+            vec!["a".into(), "b".into()],
+        ));
+        let copy_param = convert_operation_to_pyobject(input_definition)
+            .unwrap()
+            .extract::<GateDefinitionWrapper>(py)
+            .unwrap()
+            .into_py(py);
+
+        let comparison_copy =
+            bool::extract(new_op.call_method1("__eq__", (copy_param,)).unwrap()).unwrap();
+        assert!(comparison_copy);
+
+        let def_wrapper = new_op.extract::<GateDefinitionWrapper>().unwrap();
+        let new_op_diff = operation
+            .call1((
+                new_circuit(py),
+                String::from("ro"),
+                vec![2],
+                vec!["a".to_owned(), "c".to_owned()],
+            ))
+            .unwrap()
+            .downcast::<PyCell<GateDefinitionWrapper>>()
+            .unwrap();
+        let def_wrapper_diff = new_op_diff.extract::<GateDefinitionWrapper>().unwrap();
+        let helper_ne: bool = def_wrapper_diff != def_wrapper;
+        assert!(helper_ne);
+        let helper_eq: bool = def_wrapper == def_wrapper.clone();
+        assert!(helper_eq);
+
+        assert_eq!(
+            format!("{:?}", def_wrapper),
+            "GateDefinitionWrapper { internal: GateDefinition { circuit: Circuit { definitions: [], operations: [], _roqoqo_version: RoqoqoVersion }, name: \"ro\", qubits: [1], free_parameters: [\"a\", \"b\"] } }"
+        );
+    })
+}
+
+/// Test DefinitionFloat, DefinitionComplex, DefinitionUsize, DefinitionBit, InputSymbolic, name() function/input
 #[test_case(Operation::from(DefinitionFloat::new(String::from("ro"), 1, false)); "DefinitionFloat")]
 #[test_case(Operation::from(DefinitionComplex::new(String::from("ro"), 1, false)); "DefinitionComplex")]
 #[test_case(Operation::from(DefinitionUsize::new(String::from("ro"), 1, false)); "DefinitionUsize")]
@@ -343,7 +405,55 @@ fn test_pyo3_input_bit_value() {
     })
 }
 
-/// Test DefinitionFloat, DefinitionComplex, DefinitionUsize, DefinitionBit, InputSymbolic involved_qubits function
+#[cfg(feature = "unstable_operation_definition")]
+/// Test inputs for GateDefinition
+#[test]
+fn test_pyo3_gate_definition_inputs() {
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation = convert_operation_to_pyobject(Operation::from(GateDefinition::new(
+            Circuit::new(),
+            String::from("name"),
+            vec![1, 2],
+            vec!["test".into()],
+        )))
+        .unwrap();
+
+        // Test circuit()
+        let to_circuit = operation.call_method0(py, "circuit").unwrap();
+        let circuit_op = to_circuit.as_ref(py);
+        let circuit = new_circuit(py);
+        let comparison_circuit =
+            bool::extract(circuit_op.call_method1("__eq__", (circuit,)).unwrap()).unwrap();
+        assert!(comparison_circuit);
+
+        // Test name()
+        let name_op: String =
+            String::extract(operation.call_method0(py, "name").unwrap().as_ref(py)).unwrap();
+        let name_param: String = String::from("name");
+        assert_eq!(name_op, name_param);
+
+        // Test qubits()
+        let qubits: Vec<usize> = operation
+            .call_method0(py, "qubits")
+            .unwrap()
+            .as_ref(py)
+            .extract()
+            .unwrap();
+        assert_eq!(qubits, vec![1, 2]);
+
+        // Test free_parameters()
+        let free_parameters: Vec<String> = operation
+            .call_method0(py, "free_parameters")
+            .unwrap()
+            .as_ref(py)
+            .extract()
+            .unwrap();
+        assert_eq!(free_parameters, vec!["test".to_owned()]);
+    })
+}
+
+/// Test DefinitionFloat, DefinitionComplex, DefinitionUsize, DefinitionBit, InputSymbolic, involved_qubits function
 #[test_case(Operation::from(DefinitionFloat::new(String::from("ro"), 1, false)); "DefinitionFloat")]
 #[test_case(Operation::from(DefinitionComplex::new(String::from("ro"), 1, false)); "DefinitionComplex")]
 #[test_case(Operation::from(DefinitionUsize::new(String::from("ro"), 1, false)); "DefinitionUsize")]
@@ -365,6 +475,26 @@ fn test_pyo3_involved_qubits(input_definition: Operation) {
         assert_eq!(involved_op, involved_param);
     })
 }
+
+/// Test GateDefinition involved_qubits function
+#[cfg(feature = "unstable_operation_definition")]
+#[test_case(Operation::from(GateDefinition::new(Circuit::new(), String::from("ro"), vec![1], vec!["test".into()])); "GateDefinition")]
+fn test_pyo3_involved_qubits_gate_definition(input_definition: Operation) {
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation = convert_operation_to_pyobject(input_definition).unwrap();
+        let involved_op: HashSet<String> = HashSet::extract(
+            operation
+                .call_method0(py, "involved_qubits")
+                .unwrap()
+                .as_ref(py),
+        )
+        .unwrap();
+        let involved_param: HashSet<_> = HashSet::new();
+        assert_eq!(involved_op, involved_param);
+    })
+}
+
 /// Test DefinitionFloat, DefinitionComplex, DefinitionUsize, DefinitionBit format and repr functions
 #[test_case(Operation::from(DefinitionFloat::new(String::from("ro"), 1, false)), "DefinitionFloat"; "DefinitionFloat")]
 #[test_case(Operation::from(DefinitionComplex::new(String::from("ro"), 1, false)), "DefinitionComplex"; "DefinitionComplex")]
@@ -407,6 +537,30 @@ fn test_pyo3_input_symbolic_format_repr() {
     })
 }
 
+/// Test GateDefinition format and repr functions
+#[cfg(feature = "unstable_operation_definition")]
+#[test]
+fn test_pyo3_gate_definition_format_repr() {
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation = convert_operation_to_pyobject(Operation::from(GateDefinition::new(
+            Circuit::new(),
+            String::from("ro"),
+            vec![1],
+            vec!["test".into()],
+        )))
+        .unwrap();
+        let to_format = operation.call_method1(py, "__format__", ("",)).unwrap();
+        let format_op: &str = <&str>::extract(to_format.as_ref(py)).unwrap();
+        let to_repr = operation.call_method0(py, "__repr__").unwrap();
+        let repr_op: &str = <&str>::extract(to_repr.as_ref(py)).unwrap();
+        let format_repr_param: String = String::from("GateDefinition { circuit: Circuit { definitions: [], operations: [], _roqoqo_version: RoqoqoVersion }, name: \"ro\", qubits: [1], free_parameters: [\"test\"] }");
+        let comparison = format_repr_param.as_str();
+        assert_eq!(format_op, comparison);
+        assert_eq!(repr_op, comparison);
+    })
+}
+
 /// Test DefinitionFloat, DefinitionComplex, DefinitionUsize, DefinitionBit, InputSymbolic copy and deepcopy functions
 #[test_case(Operation::from(DefinitionFloat::new(String::from("ro"), 1, false)); "DefinitionFloat")]
 #[test_case(Operation::from(DefinitionComplex::new(String::from("ro"), 1, false)); "DefinitionComplex")]
@@ -441,6 +595,41 @@ fn test_pyo3_copy_deepcopy(input_definition: Operation) {
     })
 }
 
+/// Test GateDefinition copy and deepcopy functions
+#[test]
+#[cfg(feature = "unstable_operation_definition")]
+fn test_pyo3_copy_deepcopy_gate_definition() {
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation = convert_operation_to_pyobject(Operation::from(GateDefinition::new(
+            Circuit::new(),
+            "name".to_owned(),
+            vec![0],
+            vec!["param".to_owned()],
+        )))
+        .unwrap();
+        let copy_op = operation.call_method0(py, "__copy__").unwrap();
+        let deepcopy_op = operation.call_method1(py, "__deepcopy__", ("",)).unwrap();
+        let copy_deepcopy_param = operation;
+
+        let comparison_copy = bool::extract(
+            copy_op
+                .as_ref(py)
+                .call_method1("__eq__", (copy_deepcopy_param.clone(),))
+                .unwrap(),
+        )
+        .unwrap();
+        assert!(comparison_copy);
+        let comparison_deepcopy = bool::extract(
+            deepcopy_op
+                .as_ref(py)
+                .call_method1("__eq__", (copy_deepcopy_param,))
+                .unwrap(),
+        )
+        .unwrap();
+        assert!(comparison_deepcopy);
+    })
+}
 /// Test DefinitionFloat, DefinitionComplex, DefinitionUsize, DefinitionBit, InputSymbolic tags function
 #[test_case(Operation::from(DefinitionFloat::new(String::from("ro"), 1, false)), "DefinitionFloat"; "DefinitionFloat")]
 #[test_case(Operation::from(DefinitionComplex::new(String::from("ro"), 1, false)), "DefinitionComplex"; "DefinitionComplex")]
@@ -497,7 +686,34 @@ fn test_pyo3_is_parametrized(input_definition: Operation) {
     })
 }
 
-/// Test DefinitionFloat, DefinitionComplex, DefinitionUsize, DefinitionBit substitute_parameters functions
+// Test GateDefinitions's tags, hslang and is_parametrized functions
+#[cfg(feature = "unstable_operation_definition")]
+#[test_case(Operation::from(GateDefinition::new(Circuit::new(), String::from("ro"), vec![1], vec!["test".into()])); "GateDefinition")]
+fn test_pyo3_gate_definition(input_definition: Operation) {
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation = convert_operation_to_pyobject(input_definition).unwrap();
+
+        let to_tag = operation.call_method0(py, "tags").unwrap();
+        let tags_op: &Vec<&str> = &Vec::extract(to_tag.as_ref(py)).unwrap();
+        let tags_param: &[&str] = &["Operation", "Definition", "GateDefinition"];
+        assert_eq!(tags_op, tags_param);
+
+        let hqslang_op: String =
+            String::extract(operation.call_method0(py, "hqslang").unwrap().as_ref(py)).unwrap();
+        assert_eq!(hqslang_op, "GateDefinition");
+
+        assert!(!bool::extract(
+            operation
+                .call_method0(py, "is_parametrized")
+                .unwrap()
+                .as_ref(py)
+        )
+        .unwrap());
+    })
+}
+
+/// Test DefinitionFloat, DefinitionComplex, DefinitionUsize, DefinitionBit, InputSymbolic, InputBit substitute_parameters functions
 #[test_case(Operation::from(DefinitionFloat::new(String::from("ro"), 1, false)); "DefinitionFloat")]
 #[test_case(Operation::from(DefinitionComplex::new(String::from("ro"), 1, false)); "DefinitionComplex")]
 #[test_case(Operation::from(DefinitionUsize::new(String::from("ro"), 1, false)); "DefinitionUsize")]
@@ -510,6 +726,37 @@ fn test_pyo3_substitute_parameters(input_definition: Operation) {
         let operation = convert_operation_to_pyobject(input_definition).unwrap();
         let mut substitution_dict: HashMap<&str, f64> = HashMap::new();
         substitution_dict.insert("ro", 1.0);
+        let substitute_op = operation
+            .call_method1(py, "substitute_parameters", (substitution_dict,))
+            .unwrap();
+        let substitute_param = operation;
+
+        let comparison_copy = bool::extract(
+            substitute_op
+                .as_ref(py)
+                .call_method1("__eq__", (substitute_param,))
+                .unwrap(),
+        )
+        .unwrap();
+        assert!(comparison_copy);
+    })
+}
+
+/// Test GateDefinitions's substitute_parameters functions
+#[cfg(feature = "unstable_operation_definition")]
+#[test]
+fn test_pyo3_substitute_parameters_gate_definition() {
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation = convert_operation_to_pyobject(Operation::from(GateDefinition::new(
+            Circuit::new(),
+            String::from("name"),
+            vec![1],
+            vec!["test".into()],
+        )))
+        .unwrap();
+        let mut substitution_dict: HashMap<&str, f64> = HashMap::new();
+        substitution_dict.insert("name", 1.0);
         let substitute_op = operation
             .call_method1(py, "substitute_parameters", (substitution_dict,))
             .unwrap();
@@ -545,7 +792,28 @@ fn test_pyo3_substitute_parameters_error(input_operation: Operation) {
     })
 }
 
-/// Test DefinitionFloat, DefinitionComplex, DefinitionUsize, DefinitionBit remap_qubits functions
+/// Test GateDefinitions's substitute_parameters() causing an error `not-a-real-number`
+#[cfg(feature = "unstable_operation_definition")]
+#[test]
+fn test_pyo3_substitute_parameters_error_gate_definition() {
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation = convert_operation_to_pyobject(Operation::from(GateDefinition::new(
+            Circuit::new(),
+            String::from("name"),
+            vec![1],
+            vec!["test".into()],
+        )))
+        .unwrap();
+        let mut substitution_dict: HashMap<&str, &str> = HashMap::new();
+        substitution_dict.insert("name", "test");
+        let result = operation.call_method1(py, "substitute_parameters", (substitution_dict,));
+        let result_ref = result.as_ref();
+        assert!(result_ref.is_err());
+    })
+}
+
+/// Test DefinitionFloat, DefinitionComplex, DefinitionUsize, DefinitionBit, InputSymbolic, InputBit remap_qubits functions
 #[test_case(Operation::from(DefinitionFloat::new(String::from("ro"), 1, false)); "DefinitionFloat")]
 #[test_case(Operation::from(DefinitionComplex::new(String::from("ro"), 1, false)); "DefinitionComplex")]
 #[test_case(Operation::from(DefinitionUsize::new(String::from("ro"), 1, false)); "DefinitionUsize")]
@@ -563,6 +831,46 @@ fn test_pyo3_remap_qubits(input_definition: Operation) {
             .call_method1(py, "remap_qubits", (qubit_mapping,))
             .unwrap();
         let remap_param = operation;
+
+        let comparison_copy = bool::extract(
+            remap_op
+                .as_ref(py)
+                .call_method1("__eq__", (remap_param,))
+                .unwrap(),
+        )
+        .unwrap();
+        assert!(comparison_copy);
+    })
+}
+
+/// Test GateDefinitions's remap_qubits functions
+#[cfg(feature = "unstable_operation_definition")]
+#[test]
+fn test_pyo3_remap_qubits_gate_definition() {
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation = convert_operation_to_pyobject(Operation::from(GateDefinition::new(
+            Circuit::new(),
+            String::from("name"),
+            vec![0],
+            vec!["test".into()],
+        )))
+        .unwrap();
+        let remaped_operation =
+            convert_operation_to_pyobject(Operation::from(GateDefinition::new(
+                Circuit::new(),
+                String::from("name"),
+                vec![1],
+                vec!["test".into()],
+            )))
+            .unwrap();
+        let mut qubit_mapping: HashMap<usize, usize> = HashMap::new();
+        qubit_mapping.insert(0, 1);
+        qubit_mapping.insert(1, 0);
+        let remap_op = operation
+            .call_method1(py, "remap_qubits", (qubit_mapping,))
+            .unwrap();
+        let remap_param = remaped_operation;
 
         let comparison_copy = bool::extract(
             remap_op
@@ -626,6 +934,53 @@ fn test_pyo3_richcmp(definition_1: Operation, definition_2: Operation) {
     })
 }
 
+/// Test the __richcmp__ function for GateDefinitions
+#[test]
+#[cfg(feature = "unstable_operation_definition")]
+fn test_pyo3_richcmp_gate_definition() {
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation_one = convert_operation_to_pyobject(Operation::from(GateDefinition::new(
+            Circuit::new(),
+            String::from("name"),
+            vec![0],
+            vec!["test".into()],
+        )))
+        .unwrap();
+        let operation_two = convert_operation_to_pyobject(Operation::from(GateDefinition::new(
+            Circuit::new(),
+            String::from("name"),
+            vec![1],
+            vec!["testing".into()],
+        )))
+        .unwrap();
+
+        let comparison = bool::extract(
+            operation_one
+                .as_ref(py)
+                .call_method1("__eq__", (operation_two.clone(),))
+                .unwrap(),
+        )
+        .unwrap();
+        assert!(!comparison);
+
+        let comparison = bool::extract(
+            operation_one
+                .as_ref(py)
+                .call_method1("__ne__", (operation_two.clone(),))
+                .unwrap(),
+        )
+        .unwrap();
+        assert!(comparison);
+
+        let comparison = operation_one.call_method1(py, "__eq__", (vec!["fails"],));
+        assert!(comparison.is_err());
+
+        let comparison = operation_one.call_method1(py, "__ge__", (operation_two,));
+        assert!(comparison.is_err());
+    })
+}
+
 /// Test json_schema function for all define operations
 #[cfg(feature = "json_schema")]
 #[test_case(Operation::from(DefinitionFloat::new(String::from("ro"), 1, false)); "DefinitionFloat")]
@@ -662,6 +1017,39 @@ fn test_pyo3_json_schema(operation: Operation) {
             Operation::InputBit(_) => "1.1.0".to_string(),
             _ => "1.0.0".to_string(),
         };
+        let pyobject = convert_operation_to_pyobject(operation).unwrap();
+        let operation = pyobject.as_ref(py);
+
+        let schema: String =
+            String::extract(operation.call_method0("json_schema").unwrap()).unwrap();
+
+        assert_eq!(schema, rust_schema);
+
+        let current_version_string =
+            String::extract(operation.call_method0("current_version").unwrap()).unwrap();
+        let minimum_supported_version_string =
+            String::extract(operation.call_method0("min_supported_version").unwrap()).unwrap();
+
+        assert_eq!(current_version_string, ROQOQO_VERSION);
+        assert_eq!(minimum_supported_version_string, minimum_version);
+    });
+}
+
+/// Test json_schema function for GateDefinitions
+#[test]
+#[cfg(feature = "unstable_operation_definition")]
+#[cfg(feature = "json_schema")]
+fn test_pyo3_json_schema_gate_definition() {
+    let operation = Operation::from(GateDefinition::new(
+        Circuit::new(),
+        String::from("name"),
+        vec![0],
+        vec!["test".into()],
+    ));
+    let rust_schema = serde_json::to_string_pretty(&schemars::schema_for!(GateDefinition)).unwrap();
+    pyo3::prepare_freethreaded_python();
+    pyo3::Python::with_gil(|py| {
+        let minimum_version: String = "1.10.1".to_owned();
         let pyobject = convert_operation_to_pyobject(operation).unwrap();
         let operation = pyobject.as_ref(py);